♻️ [RUMF-1267] split telemetry and monitor functions
