+++ packages/core/src/browser/xhrObservable.ts
@@ -1,5 +1,5 @@
-import { monitor } from '../domain/telemetry'
 import { instrumentMethodAndCallOriginal } from '../tools/instrumentMethod'
+import { monitor } from '../tools/monitor'
 import { Observable } from '../tools/observable'
 import type { Duration, RelativeTime, ClocksState } from '../tools/timeUtils'
 import { elapsed, relativeNow, clocksNow, timeStampNow } from '../tools/timeUtils'
