+++ packages/core/src/domain/console/consoleObservable.ts
@@ -1,9 +1,9 @@
-import { callMonitored } from '../telemetry'
 import { computeStackTrace } from '../tracekit'
 import { createHandlingStack, formatErrorMessage, toStackTraceString } from '../../tools/error'
 import { mergeObservables, Observable } from '../../tools/observable'
 import { find, jsonStringify } from '../../tools/utils'
 import { ConsoleApiName } from '../../tools/display'
+import { callMonitored } from '../../tools/monitor'
 
 export interface ConsoleLog {
   message: string
