+++ packages/core/src/domain/session/sessionManager.ts
@@ -1,11 +1,11 @@
 import type { CookieOptions } from '../../browser/cookie'
 import type { Observable } from '../../tools/observable'
 import * as utils from '../../tools/utils'
-import { monitor } from '../telemetry'
 import type { Context } from '../../tools/context'
 import { ContextHistory } from '../../tools/contextHistory'
 import type { RelativeTime } from '../../tools/timeUtils'
 import { relativeNow, clocksOrigin } from '../../tools/timeUtils'
+import { monitor } from '../../tools/monitor'
 import { tryOldCookiesMigration } from './oldCookiesMigration'
 import { startSessionStore } from './sessionStore'
 import { SESSION_TIME_OUT_DELAY } from './sessionConstants'
