+++ packages/core/src/domain/telemetry/index.ts
@@ -1,14 +1,10 @@
 export {
   Telemetry,
   RawTelemetryEvent,
-  monitored,
-  monitor,
-  callMonitored,
   addTelemetryDebug,
   addTelemetryError,
   startFakeTelemetry,
   resetTelemetry,
-  setDebugMode,
   startTelemetry,
   isTelemetryReplicationAllowed,
 } from './telemetry'
